Refactoring
- Change Package name to com.soundcloud.lightcycle.integration_test
- Remove unnecessary dependency 
- Change Test SDK target to 22 to fix Robolectric bug
