--- lightcycle-integration-test/src/main/java/com/soundcloud/lightcycle/sample/basic/callback/ActivityCallback.java
@@ -1,6 +1,0 @@
-package com.soundcloud.lightcycle.sample.basic.callback;
-
-public enum ActivityCallback {
-    onNewIntent,
-    onOptionsItemSelected
-}
