+++ lightcycle-integration-test/src/test/java/com/soundcloud/lightcycle/integration_test/AppCompatActivityLoggerTest.java
@@ -1,11 +1,12 @@
-package com.soundcloud.lightcycle.sample.basic;
+package com.soundcloud.lightcycle.integration_test;
 
 import android.content.Intent;
 import android.os.Bundle;
 
 import com.google.common.truth.BooleanSubject;
-import com.soundcloud.lightcycle.sample.basic.callback.ActivityCallback;
-import com.soundcloud.lightcycle.sample.basic.callback.ActivityLifecycleCallback;
+import com.soundcloud.lightcycle.integration_test.callback.ActivityLifecycleCallback;
+import com.soundcloud.lightcycle.sample.basic.BuildConfig;
+import com.soundcloud.lightcycle.sample.basic.R;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -23,7 +24,7 @@ import java.util.List;
 import static com.google.common.truth.Truth.assertThat;
 
 @RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, sdk = 23)
+@Config(constants = BuildConfig.class, sdk = 22)
 public class AppCompatActivityLoggerTest {
     private final ActivityController<SampleAppCompatActivity> controller = Robolectric.buildActivity(SampleAppCompatActivity.class);
     private final SampleAppCompatActivity sampleAppCompatActivity = controller.get();
@@ -127,7 +128,13 @@ public class AppCompatActivityLoggerTest {
     public void onNewIntent() {
         controller.setup()
                 .newIntent(new Intent());
-        assertThat(appCompatActivityLogger.isActivityCallbackCalled(ActivityCallback.onNewIntent)).isTrue();
+
+        assertLifecycleCallbackCallIsCorrect(
+                Arrays.asList(ActivityLifecycleCallback.onCreate,
+                        ActivityLifecycleCallback.onStart,
+                        ActivityLifecycleCallback.onResume,
+                        ActivityLifecycleCallback.onNewIntent)
+        );
     }
 
     @Test
@@ -135,7 +142,13 @@ public class AppCompatActivityLoggerTest {
         controller.setup();
         ShadowActivity shadowActivity = Shadows.shadowOf(controller.get());
         shadowActivity.clickMenuItem(R.id.action_search);
-        assertThat(appCompatActivityLogger.isActivityCallbackCalled(ActivityCallback.onOptionsItemSelected)).isTrue();
+
+        assertLifecycleCallbackCallIsCorrect(
+                Arrays.asList(ActivityLifecycleCallback.onCreate,
+                        ActivityLifecycleCallback.onStart,
+                        ActivityLifecycleCallback.onResume,
+                        ActivityLifecycleCallback.onOptionsItemSelected)
+        );
     }
 
     private void assertLifecycleCallbackCallIsCorrect(List<ActivityLifecycleCallback> callbacks) {
