+++ lightcycle-integration-test/src/test/java/com/soundcloud/lightcycle/integration_test/SupportFragmentLoggerTest.java
@@ -1,15 +1,14 @@
-package com.soundcloud.lightcycle.sample.basic;
-
-import android.os.Bundle;
+package com.soundcloud.lightcycle.integration_test;
 
 import com.google.common.truth.BooleanSubject;
-import com.soundcloud.lightcycle.sample.basic.callback.FragmentLifecycleCallback;
+import com.soundcloud.lightcycle.integration_test.callback.FragmentLifecycleCallback;
+import com.soundcloud.lightcycle.sample.basic.BuildConfig;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
-import org.robolectric.util.FragmentController;
+import org.robolectric.shadows.support.v4.SupportFragmentController;
 
 import java.util.Arrays;
 import java.util.List;
@@ -17,20 +16,18 @@ import java.util.List;
 import static com.google.common.truth.Truth.assertThat;
 
 @RunWith(RobolectricTestRunner.class)
-@Config(constants = BuildConfig.class, sdk = 23)
-public class FragmentLoggerTest {
-    private final SampleFragment sampleFragment = new SampleFragment();
-    private final FragmentLogger fragmentLogger = sampleFragment.fragmentLogger;
-    private final FragmentController controller = FragmentController.of(sampleFragment);
+@Config(constants = BuildConfig.class, sdk = 22)
+public class SupportFragmentLoggerTest {
+    private final SampleSupportFragment sampleSupportFragment = new SampleSupportFragment();
+    private final SupportFragmentLogger supportFragmentLogger = sampleSupportFragment.supportFragmentLogger;
+    private final SupportFragmentController controller = SupportFragmentController.of(sampleSupportFragment);
 
     @Test
     public void onCreate() {
         controller.create();
         assertLifecycleCallbackCallIsCorrect(
                 Arrays.asList(FragmentLifecycleCallback.onAttach,
-                        FragmentLifecycleCallback.onCreate,
-                        FragmentLifecycleCallback.onViewCreated,
-                        FragmentLifecycleCallback.onActivityCreated)
+                        FragmentLifecycleCallback.onCreate)
         );
     }
 
@@ -41,9 +38,9 @@ public class FragmentLoggerTest {
         assertLifecycleCallbackCallIsCorrect(
                 Arrays.asList(FragmentLifecycleCallback.onAttach,
                         FragmentLifecycleCallback.onCreate,
+                        FragmentLifecycleCallback.onStart,
                         FragmentLifecycleCallback.onViewCreated,
-                        FragmentLifecycleCallback.onActivityCreated,
-                        FragmentLifecycleCallback.onStart)
+                        FragmentLifecycleCallback.onActivityCreated)
         );
     }
 
@@ -80,25 +77,6 @@ public class FragmentLoggerTest {
     }
 
     @Test
-    public void onSaveInstanceState() {
-        controller.create()
-                .start()
-                .resume()
-                .pause()
-                .saveInstanceState(new Bundle());
-        assertLifecycleCallbackCallIsCorrect(
-                Arrays.asList(FragmentLifecycleCallback.onAttach,
-                        FragmentLifecycleCallback.onCreate,
-                        FragmentLifecycleCallback.onViewCreated,
-                        FragmentLifecycleCallback.onActivityCreated,
-                        FragmentLifecycleCallback.onStart,
-                        FragmentLifecycleCallback.onResume,
-                        FragmentLifecycleCallback.onPause,
-                        FragmentLifecycleCallback.onSaveInstanceState)
-        );
-    }
-
-    @Test
     public void onStop() {
         controller.create()
                 .start()
@@ -131,7 +109,7 @@ public class FragmentLoggerTest {
 
     private void assertLifecycleCallbackCallIsCorrect(List<FragmentLifecycleCallback> callbacks) {
         for (FragmentLifecycleCallback fragmentLifecycleCallback : FragmentLifecycleCallback.values()) {
-            BooleanSubject subject = assertThat(fragmentLogger.isFragmentLifecycleCallbackCalled(fragmentLifecycleCallback));
+            BooleanSubject subject = assertThat(supportFragmentLogger.isFragmentLifecycleCallbackCalled(fragmentLifecycleCallback));
             if (callbacks.contains(fragmentLifecycleCallback)) {
                 subject.isTrue();
             } else {
@@ -139,4 +117,5 @@ public class FragmentLoggerTest {
             }
         }
     }
+
 }
